+++ CHANGELOG.md
@@ -4,7 +4,9 @@ CHANGELOG for Sulu Product Bundle
 * dev-develop
 
     * ENHANCEMENT Added config param 'display_recurring_prices'
-
+    * ENHANCEMENT  Added method to get products by one or more categoryIds and tags
+    * ENHANCEMENT  Added method to get products by global trade item number (gtin)
+    
 * 0.12.5 (2016-08-22)
 
     * FEATURE Added calculation of recurring prices to price-calculation util in javascript utils.